% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R
\name{custom_sensitivity}
\alias{custom_sensitivity}
\title{custom_sensitivty}
\usage{
custom_sensitivity(DM, DB, attr, algs, algParams, step, verbose, window,
  splitPercentages)
}
\arguments{
\item{DM, }{the original decision matrix}

\item{DB, }{wearhouse to store all of our iterations of data being created.}

\item{attr, }{specified attribute(s) to run sensitivty on.}

\item{algs, }{specified algorithms to run sensitivity on.}

\item{algParams, }{a list of parameters that the user wishes to customize during sensitivity analysis.}

\item{step, }{incremental step size each iteration.}

\item{verbose, }{echos print statements to understand input/output relation.}

\item{window, }{specified window size vector e.g., c(0,.5) only to allow weights to change between the given window range.}

\item{splitPercentages, }{a specified range of how each iteration will divey out the incrmented change to the other attributes. This can vary.}

\item{data, }{data.frame in read data matrix format}

\item{algParams, }{parameters to be passed around to each function in sensitivity}
}
\value{
data.frame
}
\description{
Helper function to sensitivity analysis
}
\author{
Blake Conrad \email{bmc.cs@outlook.com}
}
\keyword{data.frame,}
\keyword{ranking}
