% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R
\name{increaseAttribute}
\alias{increaseAttribute}
\title{increaseAttribute}
\usage{
increaseAttribute(DMCopyk, attr_k, step, split_step)
}
\arguments{
\item{DMCopyk, }{the original decision matrix}

\item{attr_k, }{specified attribute being examined.}

\item{step, }{incremental step size}

\item{split_step, }{how much to take away from other attributes as we increase this one}
}
\value{
data.frame
}
\description{
Helper function to sensitivity analysis
}
\author{
Blake Conrad \email{bmc.cs@outlook.com}
}
\keyword{data.frame,}
\keyword{ranking}
