% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{read.data.matrix}
\alias{read.data.matrix}
\title{read.data.matrix}
\usage{
read.data.matrix(path, header = TRUE)
}
\arguments{
\item{path, }{an absolute path to the raw decision matrix in the 'read.data.matrix' format.}

\item{header, }{Boolean, default to true, will flag the read.csv function with whether header columns are included or not.}
}
\value{
data.matrix format of a data.frame
}
\description{
This function allows you to import a csv into the correct data.matrix format for the madm package
}
\examples{

# IMPORTANT:
#  1. Abide the shape specifications listed below.
#  2. Make sure cost attributes have cost in their name. 
#  3. Make sure the weight column is explicitely labeled, 'weight'

#read.data.matrix("C:/Desktop/your.csv")

## Convert a standard import of:

#         name      x1  x2  x3Cost
# 1       weight    0.2 0.2 0.6
# 2 alternative1    0.5 0.5 0.5
# 3 alternative2    0.5 0.5 0.5
# 4 alternative3    0.5 0.5 0.5

# OR
# 
#         name      x1  x2  x3
#         benefit   +   +    -
# 1       weight    0.2 0.2 0.6
# 2 alternative1    0.5 0.5 0.5
# 3 alternative2    0.5 0.5 0.5
# 4 alternative3    0.5 0.5 0.5


# into (with rownames instead of a rownames column):

#               x1  x2  x3Cost
# weight       0.2 0.2 0.6
# alternative1 0.5 0.5 0.5
# alternative2 0.5 0.5 0.5
# alternative3 0.5 0.5 0.5

# OR

#                    x1  x2  x3
#        benefit     +   +    -
# 1       weight    0.2 0.2 0.6
# 2 alternative1    0.5 0.5 0.5
# 3 alternative2    0.5 0.5 0.5
# 4 alternative3    0.5 0.5 0.5
# Standard Decision Matrix Format (N+1)xD:

#               attribute1 attribute2 ... attributei ... attributeD
# weight     
# alternative1
# alternative2
# .
# .
# .
# alternativeN


}
\author{
Blake Conrad
}
\keyword{data.frame,}
\keyword{read.csv.}
