% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{sensitivity}
\usage{
sensitivity(data = c(), algs = c("TOPSIS", "MAUT"), algParams = c(),
  step = 0.01, window = c(), attr = c(), splitPercentages = "uniform",
  verbose = FALSE, plotLabels = FALSE)
}
\arguments{
\item{data, }{read.data.matrix(data.frame), data.frame in data matrix format.}

\item{algParams, }{list of named vectors,}

\item{step, }{automatically set to 0.01 is the steps you want an attribute to climb up or down.}

\item{window, }{double or vector w/ 1 double per attr, specifies how low and high a step size will go for each attribute relative to its 
current weight. if window exceeds 1 attr below0 or above1, the window for that attr will be void and the
 next will be assessed.}

\item{attr, }{automatically set to c() will trigger a full sensitivity accross all attributes.
If a list of valid attribuets are provided, the sensitivity will only trigger these attributes
in the sensitivity and reflect them only in the returned database.}

\item{splitPercentages, }{automatically set to uniform. This means as you incrementally increase and
decrease an attributes weight value, the percetnage given to each other attribute is equal. An example would
be with 3 attributes, we toggle attribute2 down from 0.3 to 0.2. So we lost 0.1 from attribute2, so we will take
0.1/(3-1) to give us 0.1/2==0.05, therefore attribuet1 and attribute2 will increase by 0.05 when attribute2 goes down 0.1.
The generalization to this is step/(1-N) in the algorithm.}

\item{verbose, }{default to FALSE. As with most linux command line applications, this will give you a highly detailed picture of what type of iterations are taking place under the hood.}

\item{splitPercentages, matrix}{of size n-1 by n-1, for each attr how to weigh every other attr from the inc/dec in step. each row should sum to 1.}

\item{window, }{double or vector or list, a sigle value represents how low to go for a weight on an attribute being studied, i.e weight-window. likewise for how high, weight+window. for the vector
and list situation it is nothing more than a substitute for the lower and upper bound of default, i.e., (step, 1-step)}
}
\value{
named list
}
\description{
A function to perform the sensitivity on a data.frame in read.data.matrix format.
The spirit behind this function is to interpretively take in the given parameters
of a specified read.data.matrix style data.frame and populate a database.
From this database, the user can pivot, plot, and reshape for appropriate analysis
in deciding what different weighting schemes might have looked like for the decision maker.
This function is meant to scale between various methods as well as attributes.
Custom split separation is available for each sensitivity step.
Custom step sizes are also available to allow the user to do multiple sensitivity
runs and determine what fidelity the sensitivty is needed to make an impact on the 
overarching decision.
}
\details{
Import libraries
}
\author{
Blake Conrad
}
