% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R
\name{updateDB}
\alias{updateDB}
\title{updateDB}
\usage{
updateDB(DMCopy, DB, alg, algParams, attr_i, iterid, verbose)
}
\arguments{
\item{DMCopy, }{the original decision matrix}

\item{DB, }{wearhouse to store all of our iterations of data being created.}

\item{algParams, }{parameters to be passed around to each function in sensitivity}

\item{attr_i, }{specified attribute being examined.}

\item{iterid, }{the current unique identifier to this iteration.}

\item{verbose, }{echos print statements to understand input/output relation.}

\item{algs, }{specified algorithms to run sensitivity on.}
}
\value{
data.frame
}
\description{
Helper function to sensitivity analysis
}
\author{
Blake Conrad \email{bmc.cs@outlook.com}
}
\keyword{data.frame,}
\keyword{ranking}
