% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R
\name{decreaseAttribute}
\alias{decreaseAttribute}
\title{decreaseAttribute}
\usage{
decreaseAttribute(DMCopyj, attr_j, step, split_step)
}
\arguments{
\item{DMCopyj, }{the original decision matrix}

\item{attr_j, }{specified attribute being examined.}

\item{step, }{incremental step size}

\item{split_step, }{how much to add too other attributes as we increase this one}
}
\value{
data.frame
}
\description{
Helper function to sensitivity analysis
}
\author{
Blake Conrad \email{bmc.cs@outlook.com}
}
\keyword{data.frame,}
\keyword{ranking}
