% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{MAUT}
\alias{MAUT}
\title{MAUT}
\usage{
MAUT(data = DM, algParams = c(), scales = c(), verbose = FALSE)
}
\arguments{
\item{algParams, }{default to c(), when the sensitivity function calls this function the user can pass specific algorithm parameters. These have a sanitiy check and then append in the correct parameter if correctly supplied.}

\item{scales, }{default to c(), if scales supplied and supplied correctly, each attribute will get the appropriate scaling measure (linear, exponential, or logarithmic). Defaults to all linear scales if none supplied.}

\item{read.data.matrix(data.frame), }{c()}
}
\value{
list
}
\description{
This function allows you to import a csv into the correct data.matrix format for the madm package
}
\details{
1. https://github.com/conradbm/madm/blob/master/Examples/MAUT.xls

2. https://github.com/conradbm/madm/blob/master/Examples/SAW_and_Topsis.xls

Standard Decision Matrix Format (N+1)xD:

              attribute1 attribute2 ... attributei ... attributeD
weight     
alternative1
alternative2
.
.
.
alternativeN
}
\examples{
data(maut_dm)
maut_results <- MAUT(maut_dm) #no scales specified
maut_results2 <- MAUT(maut_dm, scales=c("linear", "exponential", "logarithmic")) # all scales specified
maut_results3 <- MAUT(maut_dm, scales=c("exponential", "", "logarithmic")) # some scales specified, all others default to linear.
}
\author{
Blake Conrad \email{bmc.cs@outlook.com}
}
\keyword{data.frame,}
\keyword{ranking}
