% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{TOPSIS}
\alias{TOPSIS}
\title{TOPSIS}
\usage{
TOPSIS(data = DM, algParams = c(), verbose = FALSE)
}
\arguments{
\item{data, }{dataframe in read.data.matrix format}

\item{algParams, }{default to c(), when the sensitivity function calls this function the user can pass specific algorithm parameters. These have a sanitiy check and then append in the correct parameter if correctly supplied.}

\item{verbose, }{echos print statements to understand input/output relation.}
}
\value{
list
}
\description{
A function to perform the TOPSIS algorithm on a data.frame in read.data.matrix format
}
\details{
More theory on the following:
https://github.com/conradbm/madm/blob/master/Examples/SAW_and_Topsis.xls
}
\examples{
data(topsis_dm)
topsis_results <- TOPSIS(topsis_dm)
topsis_results$Results
}
\author{
Blake Conrad \email{bmc.cs@outlook.com}
}
\keyword{data.frame,}
\keyword{ranking}
