% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R
\name{updateDB}
\alias{updateDB}
\title{updates the data table to store new algorithm output information.
 Blake Conrad
 DMCopy, a read.data.matrix style data frame, expected to be brought in from default_sensitivity(...)
 DB, data.frame ,the 'database' or most updated dataframe object that is storing each iterations results. This function will append to it and return it.
 alg, character, specified as 'TOPSIS' or "MAUT' in version 1.0.0
algParams, list, each named item in the list is a specific parameter to the alg param specified.
 attr_i, character, a specific column name in the DMCopy which is being sensitively analyzed.
 iterid, integer, meant to keep track of indices for the DB.
 verbose, boolean, useful if TRUE for debugging.
 data.frame}
\usage{
updateDB(DMCopy, DB, alg, algParams, attr_i, iterid, verbose)
}
\description{
updates the data table to store new algorithm output information.
 Blake Conrad
 DMCopy, a read.data.matrix style data frame, expected to be brought in from default_sensitivity(...)
 DB, data.frame ,the 'database' or most updated dataframe object that is storing each iterations results. This function will append to it and return it.
 alg, character, specified as 'TOPSIS' or "MAUT' in version 1.0.0
algParams, list, each named item in the list is a specific parameter to the alg param specified.
 attr_i, character, a specific column name in the DMCopy which is being sensitively analyzed.
 iterid, integer, meant to keep track of indices for the DB.
 verbose, boolean, useful if TRUE for debugging.
 data.frame
}
